Post preregistration update
Some small changes to prep data. More importantly, we've added two new notebooks with the analysis after seeing the data and after preregistration. These changes were necessary due to mistakes in the preregistered models. This will be clarified in the paper, which will be presented as exploratory given that we changed the analysis rather substantially (although the test methods and general model structure remained the same).